converted ids from Int to Long

+++ app/src/main/java/com/flydrop2p/flydrop2p/data/local/account/AccountEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity
 data class AccountEntity(
     @PrimaryKey
-    val accountId: Int,
+    val accountId: Long,
     val profileUpdate: Long
 )
