+++ app/src/main/java/com/flydrop2p/flydrop2p/data/repository/OwnAccountLocalRepository.kt
@@ -15,7 +15,7 @@ class OwnAccountLocalRepository(private val ownAccountDataStore: DataStore<Accou
         return ownAccountDataStore.data.first()
     }
 
-    override suspend fun setAccountId(accountId: Int) {
+    override suspend fun setAccountId(accountId: Long) {
         ownAccountDataStore.updateData { it.copy(accountId = accountId) }
     }
