+++ app/src/main/java/com/flydrop2p/flydrop2p/network/model/profile/NetworkProfileResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class NetworkProfileResponse(
-    val senderId: Int,
-    val receiverId: Int,
+    val senderId: Long,
+    val receiverId: Long,
     val profile: NetworkProfile
 )
